requirements.txt
discord.py==2.3.2
pandas==2.2.2
aiohttp